requirements-agents.txt
# Alphalens Autonomous Trading System - Agent Dependencies

# Core AI & Orchestration
anthropic>=0.18.0                # Claude SDK
langgraph>=0.1.0                 # State machine orchestration
langchain>=0.1.0                 # LangChain base
langchain-anthropic>=0.1.0       # Claude integration for LangChain

# Memory & State Management
redis>=5.0.0                     # In-memory cache
psycopg2-binary>=2.9.0          # PostgreSQL connector
sqlalchemy>=2.0.0                # ORM for database
weaviate-client>=3.25.0         # Vector database for semantic memory (alternative: pinecone-client)

# Data & Analysis
pandas>=0.18.0
numpy>=1.9.1
scipy>=0.14.0

# API & Background Tasks
fastapi>=0.104.0                 # API server
uvicorn>=0.24.0                  # ASGI server
celery>=5.3.0                    # Background task queue
redis[hiredis]>=5.0.0           # Celery broker

# Monitoring & Observability
prometheus-client>=0.19.0        # Metrics collection
sentry-sdk>=1.38.0              # Error tracking
langsmith>=0.0.69               # LangGraph tracing

# Trading & Backtesting
pyfolio>=0.9.2                  # Portfolio analysis (Quantopian)
empyrical>=0.5.0                # Financial metrics
zipline-reloaded>=2.4.0         # Backtesting framework (optional)

# Broker Integrations
alpaca-py>=0.14.0               # Alpaca broker API
ib_insync>=0.9.86               # Interactive Brokers (optional)

# Data Feeds
yfinance>=0.2.28                # Yahoo Finance data (free)
alpaca-py>=0.14.0               # Alpaca data feed (included above)

# Dashboard & Visualization
streamlit>=1.28.0               # Streamlit dashboard
plotly>=5.17.0                  # Interactive charts
requests>=2.31.0                # HTTP requests for API client

# Utilities
python-dotenv>=1.0.0            # Environment variables
pydantic>=2.0.0                 # Data validation
tenacity>=8.2.0                 # Retry logic
loguru>=0.7.0                   # Better logging

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
black>=23.0.0                    # Code formatting
ruff>=0.1.0                      # Fast linting
